+++ src/win/canvasimpld2d.h
@@ -148,8 +148,8 @@ namespace k_canvas
 
             void GetFontMetrics(const kFontBase *font, kFontMetrics &metrics) override;
             void GetGlyphMetrics(const kFontBase *font, size_t first, size_t last, kGlyphMetrics *metrics) override;
-            kSize TextSize(const char *text, int count, const kFontBase *font) override;
-            void Text(const kPoint &p, const char *text, int count, const kFontBase *font, const kBrushBase *brush, kTextOrigin origin) override;
+            kSize TextSize(const char *text, size_t count, const kFontBase *font) override;
+            void Text(const kPoint &p, const char *text, size_t count, const kFontBase *font, const kBrushBase *brush, kTextOrigin origin) override;
 
             void BeginClippedDrawingByMask(const kBitmapImpl *mask, const kTransform &transform, kExtendType xextend, kExtendType yextend) override;
             void BeginClippedDrawingByPath(const kPathImpl *clip, const kTransform &transform) override;
